+++ examples/mouse.rs
@@ -4,7 +4,9 @@ use std::time::Duration;
 
 fn main() {
     let wait_time = Duration::from_secs(2);
+
     let mut enigo = Enigo::new();
+    // let mut enigo = enigo::XConnection::default();
 
     thread::sleep(Duration::from_secs(4));
     println!("screen dimensions: {:?}", enigo.main_display_size());
@@ -14,12 +16,14 @@ fn main() {
 
     enigo.mouse_move_to(500, 200);
     thread::sleep(wait_time);
+    println!("mouse location: {:?}", enigo.mouse_location());
 
     enigo.mouse_down(MouseButton::Left);
     thread::sleep(wait_time);
 
     enigo.mouse_move_relative(100, 100);
     thread::sleep(wait_time);
+    println!("mouse location: {:?}", enigo.mouse_location());
 
     enigo.mouse_up(MouseButton::Left);
     thread::sleep(wait_time);
@@ -29,12 +33,15 @@ fn main() {
 
     enigo.mouse_scroll_x(2);
     thread::sleep(wait_time);
+    println!("mouse location: {:?}", enigo.mouse_location());
 
     enigo.mouse_scroll_x(-2);
     thread::sleep(wait_time);
+    println!("mouse location: {:?}", enigo.mouse_location());
 
     enigo.mouse_scroll_y(2);
     thread::sleep(wait_time);
+    println!("mouse location: {:?}", enigo.mouse_location());
 
     enigo.mouse_scroll_y(-2);
     thread::sleep(wait_time);